+++ src/PhpWord/Element/AbstractElement.php
@@ -101,6 +101,13 @@ abstract class AbstractElement
     private $parent;
 
     /**
+     * changed element info
+     *
+     * @var TrackChange
+     */
+    private $trackChange;
+
+    /**
      * Parent container type
      *
      * @var string
@@ -436,6 +443,38 @@ abstract class AbstractElement
         }
 
         return $style;
+    }
+
+    /**
+     * Sets the trackChange information
+     *
+     * @param TrackChange $trackChange
+     */
+    public function setTrackChange(TrackChange $trackChange)
+    {
+        $this->trackChange = $trackChange;
+    }
+
+    /**
+     * Gets the trackChange information
+     *
+     * @return TrackChange
+     */
+    public function getTrackChange()
+    {
+        return $this->trackChange;
+    }
+
+    /**
+     * Set changed
+     *
+     * @param string $type INSERTED|DELETED
+     * @param string $author
+     * @param null|int|\DateTime $date allways in UTC
+     */
+    public function setChangeInfo($type, $author, $date = null)
+    {
+        $this->trackChange = new TrackChange($type, $author, $date);
     }
 
     /**