+++ CHANGELOG.md
@@ -9,6 +9,7 @@ v0.15.0 (?? ??? 2018)
 - Parsing of "align" HTML attribute - @troosan #1231
 - Parse formatting inside HTML lists - @troosan @samimussbach #1239 #945 #1215 #508
 - Parsing of CSS `direction` instruction, HTML `lang` attribute, formatting inside table cell - @troosan #1273 #1252 #1254
+- Add support for Track changes @Cip @troosan #354 #1262
 
 ### Fixed
 - fix reading of docx default style - @troosan #1238